Extract added-field classification from TransformWorkerProgram
Move the listed members out of TransformWorkerProgram without changing their bodies.
